ai_llm_module/requirements.txt
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0
flask==2.0.1
pandas==1.3.3
anthropic==0.34.2 